Couple of changes, cleaned up some EventController code (don't think I broke anything), created a unique code that gets generated when a user attends and event that gets stored in the rsvp table, printed on the user's ticket and printed on the all attendees admin page. Also added a function that prints a barcode onto the ticket that refers to this text. You need to run sudo apt-get install php5-gd to make the image creation work, and refresh you DB because of the DB changes.

+++ app/Http/Controllers/EventController.php
@@ -142,7 +142,6 @@ class EventController extends Controller
                 ->join('users', 'users.id', '=', 'rsvp.userid')
                 ->select('events.*')
                 ->where('rsvp.userid', '=', \Auth::user()->id)
-                ->where('rsvp.deleted_at', NULL)
                 ->distinct()
                 ->get();
 
@@ -153,7 +152,11 @@ class EventController extends Controller
     // Database insertion links user to event
     public function attendEvent($id) 
     {
-        \DB::table('rsvp')->insert(['userid' => \Auth::user()->id, 'eventid' => $id]);
+        do {
+            $code = str_random(10);
+        } while (Rsvp::where("code", $code)->where('eventid', $id)->first() instanceof Rsvp);
+        Rsvp::insert(['userid' => \Auth::user()->id, 'eventid' => $id, 'code' => $code]);
+        
         return redirect('dash');     
     }
 
@@ -161,9 +164,8 @@ class EventController extends Controller
     //Database deletion removes link between user and event
     public function unattendEvent($id) 
     {
-        \DB::table('rsvp')->where(['userid' => \Auth::user()->id, 'eventid' => $id])->delete();
-        Rsvp::where('eventid',$id)->delete();
-                return redirect('dash');     
+        Rsvp::where(['userid' => \Auth::user()->id, 'eventid' => $id])->delete();
+        return redirect('/events');     
     }
